refactor code to use Vue syntax and move ajax to api endpoint
