--- app/cdash/public/css/d3.dependencyedgebundling.css
@@ -1,113 +1,0 @@
-.node {
-  /* font: 400 14px "Helvetica Neue", Helvetica, Arial, sans-serif; */
-  font-weight: 400;
-  font-size: 13px;
-  fill: #888;
-  font-family: 'Open Sans', sans-serif;
-  opacity: 0.8;
-}
-
-.node:hover {
-  fill: #000;
-  cursor: hand;
-  cursor: pointer;
-}
-
-.link {
-  stroke: steelblue;
-  stroke-opacity: .4;
-  fill: none;
-  pointer-events: none;
-  opacity: 0.4;
-}
-
-.node:hover,
-.node--source,
-.node--target {
-  fill: #000;
-  font-weight: 700;
-  opacity: 1;
-}
-
-.node--source {
-  fill: #2ca02c;
-}
-
-.node--target {
-  fill: #d62728;
-}
-
-.link--source,
-.link--target {
-  stroke-opacity: 1;
-  stroke-width: 2px;
-  opacity: 1;
-}
-
-.link--source {
-  stroke: #d62728;
-}
-
-.link--target {
-  stroke: #2ca02c;
-}
-
-div.tooltip {
-    position: absolute; /* reference for measurement */
-    text-align: left;
-    pointer-events: none; /* 'none' tells the mouse to ignore the rectangle */
-    background: #E0E0E0;
-    padding: 10px;
-    border: 1px solid #D5D5D5;
-    font-family: arial,helvetica,sans-serif;
-    position: absolute;
-    font-size: 1.1em;
-    color: #333;
-    padding: 10px;
-    border-radius: 3px;
-    background: rgba(255,255,255,0.9);
-    color: #000;
-    box-shadow: 0 1px 5px rgba(0,0,0,0.4);
-    -moz-box-shadow: 0 1px 5px rgba(0,0,0,0.4);
-    border:1px solid rgba(200,200,200,0.85);
-    z-index: 10000;
-}
-
-div.tooltipTail {
-    position: absolute;
-    left:-7px;
-    top: 12px;
-    width: 7px;
-    height: 13px;
-    background: url("../img/tail_white.png") 50% 0%;
-}
-
-div.toolTipBody {
-    position:absolute;
-    height:100px;
-    width:230px;
-}
-
-#toolTip .header {
-    text-align: left;
-    font-size: 14px;
-    margin-bottom: 2px;
-    color:#000;
-    font-weight: 700;
-}
-
-div.header-rule{
-    height:1px;
-    margin:1px auto 3px;
-    margin-top:7px;
-    margin-bottom:7px;
-    background:#ddd;
-    width:125px;
-}
-
-div.header1{
-    text-align: left;
-    font-size: 12px;
-    margin-bottom: 2px;
-    color:black;
-}
