+++ app/cdash/public/views/viewSubProjects.html
@@ -121,6 +121,6 @@
 
       <br/>
 
-      <a ng-href="viewSubProjectDependenciesGraph.php?{{cdash.linkparams}}">
+      <a ng-href="viewSubProjectDependenciesGraph?{{cdash.linkparams}}">
         [SubProject Dependencies]
       </a>
