+++ resources/js/components/SubProjectDependencies.vue
@@ -1,33 +1,48 @@
 <template>
   <div style="top:10px; left:20px; overflow:hidden;">
-    <label style="font-size:1.2em;"><b>SubProject Dependencies Graph</b></label>
-    <label for="selectedsort" style="margin-left:80px; font-size:.95em;">Sorted by:</label>
-    <select id="selectedsort">
-      <option value="0" selected="selected">subproject name</option>
-      <option value="1">subproject id</option>
-    </select>
-    <button onclick="download_svg()" style="float:right; width:200px; margin-right:30px">Export as svg file</button>
+    <label style="font-size:1.2em;">
+      <b>SubProject Dependencies Graph</b>
+    </label>
+    <span style="float:right;">
+      <label for="selectedsort" style="margin-right:10px; font-size:.95em;">Sorted by:</label>
+      <select id="selectedsort" @change="apply_sorting($event)">
+        <option value="0" selected="selected">subproject name</option>
+        <option value="1">subproject id</option>
+      </select>
+    </span>
   </div>
   <div class="hint" style="top:20px; left:20px; font-size:0.9em; width:350px; color:#999;">
     This circle plot captures the interrelationships among subgroups. Mouse over any of the subgroup in this graph to see incoming links (dependents) in green and the outgoing links (dependencies) in red.
   </div>
-  <div id="chart_placeholder" style="left:150px; top:-60px; text-align:center;">
+  <div style="text-align:center;">
+    <img v-if="graphLoading" :src="$baseURL + '/img/loading.gif'">
+    <div id="chart_placeholder"></div>
   </div>
   <!-- Tooltip -->
   <div id="toolTip" class="tooltip" style="opacity:0;">
     <div id="header1" class="header"></div>
     <div id="dependency" style="color:#d62728;"></div>
     <div id="dependents" style="color:#2ca02c;"></div>
-    <div  class="tooltipTail"></div>
+    <div class="tooltipTail"></div>
   </div>
-  <link href="//fonts.googleapis.com/css?family=Open+Sans:400,700|Roboto:400,700" rel="stylesheet" type="text/css"/>
-  <link href="{{ asset('css/d3.dependencyedgebundling.css') }}" rel="stylesheet" type="text/css"/>
 </template>
+
 <script>
 import ApiLoader from './shared/ApiLoader';
-import * from '../../../app/cdash/public/js/d3.dependencyedgebundling.js';
+import DependencyEdgeBundling from './shared/DependencyEdgeBundling.js';
 export default {
   name: 'SubProjectDependencies',
+
+  props: {
+    projectName: {
+      type: String,
+      required: true,
+    },
+    date: {
+      type: String,
+      required: true,
+    },
+  },
 
   data () {
     return {
@@ -35,38 +50,27 @@ export default {
       cdash: {},
       loading: true,
       errored: false,
+      depData: {},
+      graphLoading: true,
+      chart: {},
     }
   },
 
   mounted () {
-    ApiLoader.loadPageData(this, '/api/v1/???'); // TODO
-    this.plot_subdependencies();
+    ApiLoader.loadPageData(this, `/api/v1/getSubProjectDependencies.php?project=${this.projectName}&date=${this.date}`);
   },
 
   methods: {
-    plot_subdependencies() {
-      var chart = d3.chart.dependencyedgebundling();
-      chart.mouseOvered(mouseOvered).mouseOuted(mouseOuted);
-      var dataroot;
-      function sort_by_name (a, b) {
-        if (a.name < b.name) {
-          return -1;
-        }
-        if (a.name > b.name) {
-          return 1;
-        }
-        return 0;
-      }
+    postSetup(response) {
+      this.depData = response.data.dependencies;
+      this.chart = DependencyEdgeBundling.initChart()
+      this.plot_subdependencies();
+    },
 
-      function sort_by_id (a, b) {
-        if (a.id < b.id) {
-          return -1;
-        }
-        if (a.id > b.id) {
-          return 1;
-        }
-          return 0;
-      }
+    plot_subdependencies() {
+
+      const vm = this;
+      vm.chart.mouseOvered(mouseOvered).mouseOuted(mouseOuted);
 
       function mouseOvered(d) {
         var header1Text = "Name: " + d.key;
@@ -102,36 +106,161 @@ export default {
         d3.select("#toolTip").style("opacity", "0");
       }
 
-      function resetDepView() {
-        d3.select('#chart_placeholder svg').remove();
-        d3.select('#chart_placeholder')
-          .datum(dataroot)
-          .call(chart);
+      vm.apply_sorting({ target: { value: 0 } }); // load the graph for the first time
+    },
+
+    // replot the graph after change (or on initial load)
+    resetDepView() {
+      const vm = this;
+      vm.graphLoading = true;
+      d3.select('#chart_placeholder svg').remove();
+      d3.select('#chart_placeholder')
+        .datum(vm.depData)
+        .call(vm.chart);
+      vm.graphLoading = false;
+    },
+
+    // event listener to resort graph according to selection field
+    apply_sorting(e) {
+      const vm = this;
+      const selected = e.target.value;
+      if (parseInt(selected) === 1) {
+        vm.depData.sort(sort_by_id);
+      } else if (parseInt(selected) === 0) {
+        vm.depData.sort(sort_by_name);
       }
-      $(function(){
-        $('#selectedsort').on("change", function(e) {
-          selected = $(this).val();
-          if (parseInt(selected) === 1) {
-            dataroot.sort(sort_by_id);
-          } else if (parseInt(selected) === 0) {
-            dataroot.sort(sort_by_name);
-          }
-          resetDepView(dataroot);
-        });
-        var ajaxlink = "{{ url('ajax/getsubprojectdependencies.php') }}?project={{ urlencode($project->Name) }}";
-        d3.json(ajaxlink, function(error, classes) {
-          if (error){
-            errormsg = "json error " + error + " data: " + classes;
-            console.log(errormsg);
-            document.write(errormsg);
-            return;
-          }
-          dataroot = classes;
-          dataroot.sort(sort_by_name);
-          resetDepView(dataroot);
-        });
-      });
+      vm.resetDepView();
+
+      function sort_by_name (a, b) {
+        if (a.name < b.name) {
+          return -1;
+        }
+        if (a.name > b.name) {
+          return 1;
+        }
+        return 0;
+      }
+
+      function sort_by_id (a, b) {
+        if (a.id < b.id) {
+          return -1;
+        }
+        if (a.id > b.id) {
+          return 1;
+        }
+          return 0;
+      }
+
     },
   },
 };
 </script>
+
+<style>
+/* these can't be under a scoped style tag because they
+   target elements generated by d3 that Vue doesn't track */
+.node {
+  font-weight: 400;
+  font-size: 13px;
+  fill: #888;
+  opacity: 0.8;
+}
+
+.node:hover {
+  fill: #000;
+  cursor: hand;
+  cursor: pointer;
+}
+
+.link {
+  stroke: steelblue;
+  stroke-opacity: .4;
+  fill: none;
+  pointer-events: none;
+  opacity: 0.4;
+}
+
+.node:hover,
+.node--source,
+.node--target {
+  fill: #000;
+  font-weight: 700;
+  opacity: 1;
+}
+
+.node--source {
+  fill: #2ca02c;
+}
+
+.node--target {
+  fill: #d62728;
+}
+
+.link--source,
+.link--target {
+  stroke-opacity: 1;
+  stroke-width: 2px;
+  opacity: 1;
+}
+
+.link--source {
+  stroke: #d62728;
+}
+
+.link--target {
+  stroke: #2ca02c;
+}
+
+div.tooltip {
+    position: absolute; /* reference for measurement */
+    text-align: left;
+    pointer-events: none; /* 'none' tells the mouse to ignore the rectangle */
+    background: #E0E0E0;
+    padding: 10px;
+    border: 1px solid #D5D5D5;
+    font-family: arial,helvetica,sans-serif;
+    position: absolute;
+    font-size: 1.1em;
+    color: #333;
+    padding: 10px;
+    border-radius: 3px;
+    background: rgba(255,255,255,0.9);
+    color: #000;
+    box-shadow: 0 1px 5px rgba(0,0,0,0.4);
+    -moz-box-shadow: 0 1px 5px rgba(0,0,0,0.4);
+    border:1px solid rgba(200,200,200,0.85);
+    z-index: 10000;
+}
+</style>
+
+<style scoped>
+div.tooltipTail {
+    position: absolute;
+    left:-7px;
+    top: 12px;
+    width: 7px;
+    height: 13px;
+    background: url("/img/tail_white.png") 50% 0%;
+}
+
+div.toolTipBody {
+    position:absolute;
+    height:100px;
+    width:230px;
+}
+
+#toolTip .header {
+    text-align: left;
+    font-size: 14px;
+    margin-bottom: 2px;
+    color:#000;
+    font-weight: 700;
+}
+
+div.header1{
+    text-align: left;
+    font-size: 12px;
+    margin-bottom: 2px;
+    color:black;
+}
+</style>
